backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
Pillow>=10.0.0
numpy>=1.24.0,<3.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
requests>=2.31.0
opencv-python-headless>=4.8.0
urllib3>=2.0.0

